backend/ai-api/requirements.txt
rembg
fastapi
uvicorn
pillow
python-multipart
onnxruntime # Use onnxruntime-gpu if you have a GPU
